+++ src/main/java/subway/application/StationService.java
@@ -1,7 +1,7 @@
 package subway.application;
 
 import org.springframework.stereotype.Service;
-import subway.dao.StationDao;
+import subway.persistence.StationDao;
 import subway.domain.Station;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
