+++ src/main/java/subway/persistence/dto/LineDto.java
@@ -1,4 +1,4 @@
-package subway.dao.dto;
+package subway.persistence.dto;
 
 public class LineDto {
     private Long id;
