+++ src/main/java/subway/persistence/dto/SectionDto.java
@@ -1,4 +1,4 @@
-package subway.dao.dto;
+package subway.persistence.dto;
 
 public class SectionDto {
